--- a/.azure/linux-stack.yml
+++ b/.azure/linux-stack.yml
@@ -32,7 +32,9 @@ jobs:
       ./fmt.sh
     displayName: "HLint via ./fmt.sh"
   - bash: |
-      sudo apt-get install -y g++ gcc libc6-dev libffi-dev libgmp-dev make zlib1g-dev
+      sudo add-apt-repository ppa:hvr/ghc
+      sudo apt-get update
+      sudo apt-get install -y g++ gcc libc6-dev libffi-dev libgmp-dev make zlib1g-dev cabal-install-3.2
       if ! which stack >/dev/null 2>&1; then
          curl -sSL https://get.haskellstack.org/ | sh
       fi