--- a/ghcide.cabal
+++ b/ghcide.cabal
@@ -15,7 +15,13 @@ homepage:           https://github.com/digital-asset/ghcide#readme
 bug-reports:        https://github.com/digital-asset/ghcide/issues
 tested-with:        GHC==8.6.5
 extra-source-files: include/ghc-api-version.h README.md CHANGELOG.md
-                    test/data/GotoHover.hs
+                    test/data/hover/*.hs
+                    test/data/multi/cabal.project
+                    test/data/multi/hie.yaml
+                    test/data/multi/a/a.cabal
+                    test/data/multi/a/*.hs
+                    test/data/multi/b/b.cabal
+                    test/data/multi/b/*.hs
 
 source-repository head
     type:     git