--- a/src/Development/IDE/GHC/Compat.hs
+++ b/src/Development/IDE/GHC/Compat.hs
@@ -306,8 +306,4 @@ getConArgs = GHC.getConDetails
 #endif
 
 getPackageName :: DynFlags -> Module.InstalledUnitId -> Maybe PackageName
-#if MIN_GHC_API_VERSION(8,10,0)
-getPackageName dfs i = getPackageName <$> lookupPackage dfs (DefUnitId i)
-#else
 getPackageName dfs i = packageName <$> lookupPackage dfs (Module.DefiniteUnitId (Module.DefUnitId i))
-#endif