--- a/test/data/multi/a/A.hs
+++ b/test/data/multi/a/A.hs
@@ -1,0 +1,3 @@
+module A where
+
+foo = ()