--- a/test/data/multi/a/A.hs
+++ b/test/data/multi/a/A.hs
@@ -1,3 +1,3 @@
-module A where
+module A(foo) where
 
 foo = ()