--- a/test/data/multi/a/a.cabal
+++ b/test/data/multi/a/a.cabal
@@ -1,0 +1,9 @@
+name: a
+version: 1.0.0
+build-type: Simple
+cabal-version: >= 1.2
+
+library
+  build-depends: base
+  exposed-modules: A
+  hs-source-dirs: .