--- a/test/data/multi/b/B.hs
+++ b/test/data/multi/b/B.hs
@@ -1,3 +1,3 @@
-module B where
+module B(module B) where
 import A
 qux = foo