--- a/test/data/multi/b/b.cabal
+++ b/test/data/multi/b/b.cabal
@@ -1,0 +1,9 @@
+name: b
+version: 1.0.0
+build-type: Simple
+cabal-version: >= 1.2
+
+library
+  build-depends: base, a
+  exposed-modules: B
+  hs-source-dirs: .