--- a/test/data/multi/cabal.project
+++ b/test/data/multi/cabal.project
@@ -1,0 +1,1 @@
+packages: a b