--- a/test/data/multi/hie.yaml
+++ b/test/data/multi/hie.yaml
@@ -1,0 +1,6 @@
+cradle:
+  cabal:
+    - path: "./a"
+      component: "lib:a"
+    - path: "./b"
+      component: "lib:b"