… .azure/windows-stack.yml
… .ghci
… README.md
… exe/RuleTypes.hs
… exe/Rules.hs
… exe/Util.hs
… hie.yaml
… src/Development/IDE/Types/Options.hs
… test/data/hover/Bar.hs
… test/data/hover/Foo.hs
… test/data/hover/GotoHover.hs
… test/exe/Main.hs