@runeksvendsen 2020-04-07
I seem to experience an intermittent error using this branch of `ghcide` in a project of mine, in which a test and an executable depend on the library. The error is that the library in question ( `rule-lang`) cannot be found:

```
ghcide: <command line>: cannot satisfy -package rule-lang-0.1.0.0
```

After outputting the above `ghcide` hangs (i.e. the executable never exits, but simply hangs after printing the above line).

I made a commit of the code that exhibits the above symptoms here: https://github.com/runeksvendsen/rule-lang/commit/6bdea22bce58e1efa1c8ae4f60a3e28ebac35d6b.

I'm wondering if there's some non-determinism in the way multiple components are loaded, such that if an executable or test (which depend on the library) is loaded before the library then ghcide will complain that the library can't be found.

Full log output of running `ghcide` executable :

```
ghcide version: 0.1.0 (GHC: 8.6.5) (PATH: /Users/runesvendsen/.local/bin/ghcide) (GIT hash: c43a078c829be65d002c6ebcd3b1c9f2f534ca43)
Ghcide setup tester in /Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs.
Report bugs at https://github.com/digital-asset/ghcide/issues

Step 1/6: Finding files to test in /Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs
Found 17 files

Step 2/6: Looking for hie.yaml files that control setup
Found 1 cradle

Step 3/6: Initializing the IDE

Step 4/6: Type checking the files
Shelling out to cabal "/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/src/Absyn.hs"
> rule-lang> initial-build-steps (lib + exe)
> Configuring GHCi with the following packages: rule-lang
> /Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/install/x86_64-osx/94377dc6990ad6b243b760bbfb93aab30063cf768abe840214a74b980aac577e/8.6.5/pkgdb:/Users/runesvendsen/.stack/snapshots/x86_64-osx/94377dc6990ad6b243b760bbfb93aab30063cf768abe840214a74b980aac577e/8.6.5/pkgdb:/Users/runesvendsen/.stack/programs/x86_64-osx/ghc-8.6.5/lib/ghc-8.6.5/package.conf.d
ComponentOptions {componentOptions = ["-i","-odir=/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/odir","-hidir=/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/odir","-hide-all-packages","-XOverloadedStrings","-XStandaloneDeriving","-XDeriveGeneric","-XNoImplicitPrelude","-XFlexibleContexts","-i/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/dist/x86_64-osx/Cabal-2.4.0.1/build","-i/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/src","-i/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/dist/x86_64-osx/Cabal-2.4.0.1/build/autogen","-i/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/dist/x86_64-osx/Cabal-2.4.0.1/build/global-autogen","-stubdir=/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/dist/x86_64-osx/Cabal-2.4.0.1/build","-package-id=aeson-1.4.4.0-8dIyFiOKltcCLD7t2J0OHI","-package-id=base-4.12.0.0","-package-id=containers-0.6.0.1","-package-id=errors-2.3.0-GKWdqDKmaPrGy1XArjac2y","-package-id=hashable-1.2.7.0-2SI038axTEd7AEZJ275kpi","-package-id=megaparsec-7.0.5-9VrJ76Lt7nJ5Iji4PA7A8S","-package-id=parsers-0.12.10-GWrQJtO53B83vKycLKDMVC","-package-id=protolude-0.2.3-6EKDEuuh3ZRFm0XjK4luPw","-package-id=text-1.2.3.1","-package-id=transformers-0.5.6.2","-package-id=unordered-containers-0.2.10.0-LgoTL3wbBEY5bZIDJiyxW4","-Wall","-optP-include","-optP/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/ghci/168fd77d/cabal_macros.h","-ghci-script=/private/var/folders/dm/d2y4x8mx50n__pwvsmpj_j_r0000gn/T/haskell-stack-ghci/fda1367d/ghci-script","-package-db","/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/install/x86_64-osx/94377dc6990ad6b243b760bbfb93aab30063cf768abe840214a74b980aac577e/8.6.5/pkgdb","-package-db","/Users/runesvendsen/.stack/snapshots/x86_64-osx/94377dc6990ad6b243b760bbfb93aab30063cf768abe840214a74b980aac577e/8.6.5/pkgdb","-package-db","/Users/runesvendsen/.stack/programs/x86_64-osx/ghc-8.6.5/lib/ghc-8.6.5/package.conf.d"], componentDependencies = ["rule-lang.cabal","package.yaml","stack.yaml"]}
"Making new HscEnv[main]"
(HscEnvEq 1,fromList [("package.yaml",Just 2020-03-31 08:05:45.386518558 UTC),("rule-lang.cabal",Just 2020-04-07 08:50:32.00053941 UTC),("stack.yaml",Just 2020-03-17 16:08:13.409014676 UTC)])
Shelling out to cabal "/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/test/Spec/Parse.hs"
> Using main module: 1. Package `rule-lang' component rule-lang:test:rule-lang-test with main-is file: /Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/test/Spec.hs
> rule-lang> initial-build-steps (lib + exe)
> The following GHC options are incompatible with GHCi and have not been passed to it: -O2
> Configuring GHCi with the following packages: rule-lang
> /Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/install/x86_64-osx/94377dc6990ad6b243b760bbfb93aab30063cf768abe840214a74b980aac577e/8.6.5/pkgdb:/Users/runesvendsen/.stack/snapshots/x86_64-osx/94377dc6990ad6b243b760bbfb93aab30063cf768abe840214a74b980aac577e/8.6.5/pkgdb:/Users/runesvendsen/.stack/programs/x86_64-osx/ghc-8.6.5/lib/ghc-8.6.5/package.conf.d
ComponentOptions {componentOptions = ["-i","-odir=/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/odir","-hidir=/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/odir","-hide-all-packages","-XOverloadedStrings","-XStandaloneDeriving","-XDeriveGeneric","-XNoImplicitPrelude","-XFlexibleContexts","-i/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/dist/x86_64-osx/Cabal-2.4.0.1/build/rule-lang-test","-i/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/test","-i/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/dist/x86_64-osx/Cabal-2.4.0.1/build/rule-lang-test/autogen","-i/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/dist/x86_64-osx/Cabal-2.4.0.1/build/global-autogen","-i/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/dist/x86_64-osx/Cabal-2.4.0.1/build/rule-lang-test/rule-lang-test-tmp","-stubdir=/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/dist/x86_64-osx/Cabal-2.4.0.1/build","-package-id=base-4.12.0.0","-package-id=hspec-2.7.1-BxmpM3H5M4w4Y2ViaUKRtb","-package-id=hspec-expectations-pretty-diff-0.7.2.4-DNpNgJ1MOyu63xg9pNKuPu","-package-id=hspec-smallcheck-0.5.2-7Iq7U1gcFRNBmJABwPLv5E","-package-id=megaparsec-7.0.5-9VrJ76Lt7nJ5Iji4PA7A8S","-package=rule-lang-0.1.0.0","-package-id=smallcheck-1.1.5-8zXm09BFgY5BYxznod0SEL","-package-id=tasty-1.2.3-hTb5MwkGozC98ZFlavggW","-package-id=tasty-smallcheck-0.8.1-1OLucf7iboX6WjE9zcAsBp","-package-id=text-1.2.3.1","-Wall","-optP-include","-optP/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/ghci/33d2892b/cabal_macros.h","-ghci-script=/private/var/folders/dm/d2y4x8mx50n__pwvsmpj_j_r0000gn/T/haskell-stack-ghci/29833fcc/ghci-script","-package-db","/Users/runesvendsen/Documents/ITU/06BSc/code/RuleLangHs/.stack-work/install/x86_64-osx/94377dc6990ad6b243b760bbfb93aab30063cf768abe840214a74b980aac577e/8.6.5/pkgdb","-package-db","/Users/runesvendsen/.stack/snapshots/x86_64-osx/94377dc6990ad6b243b760bbfb93aab30063cf768abe840214a74b980aac577e/8.6.5/pkgdb","-package-db","/Users/runesvendsen/.stack/programs/x86_64-osx/ghc-8.6.5/lib/ghc-8.6.5/package.conf.d"], componentDependencies = ["rule-lang.cabal","package.yaml","stack.yaml"]}
ghcide: <command line>: cannot satisfy -package rule-lang-0.1.0.0
    (use -v for more information)
^C
```