@mpickering 2020-04-07
@runeksvendsen I tested your branch using cabal and it worked for me (after I fixed the build errors in the library component).