@mpickering 2020-04-07
I noticed the logic to ensure that multiple initialisations do not happen concurrently is not currently correct.