@pepeiborra 2020-05-30
> @pepeiborra I really do not want to do that given how long this review cycle has already gone on for. I am now turning my attention to upstreaming other patches from my branch so we can get Zubin's hiedb patches merged which will alter any timing measurements anyway.

I am very keen in seeing this PR get merged too, as I do have 2 PRs which have been open for weeks waiting for this one to get merged first. But given how cheap it is to run the benchmark suite, and the benefit of understanding what the performance impact is, if any, I don't see any reason not to do it. This is what I get in my home desktop PC (2011 quad-core Mac) : 
```
Benchmark BEFORE:

TOTAL hover = 2.57s (10 repetitions)
TOTAL getDefinition = 2.67s (10 repetitions)
TOTAL documentSymbols = 1.52s (100 repetitions)
TOTAL documentSymbols after edit = 2.92s (100 repetitions)
TOTAL completions after edit = 4.51s (10 repetitions)
TOTAL code actions = 2.47s (10 repetitions)
TOTAL code actions after edit = 3.59s (10 repetitions)
Benchmark ghcide-bench: FINISH

Benchmark AFTER:

TOTAL hover = 3.35s (10 repetitions)
TOTAL getDefinition = 3.37s (10 repetitions)
TOTAL documentSymbols = 1.39s (100 repetitions)
TOTAL documentSymbols after edit = 3.02s (100 repetitions)
TOTAL completions after edit = 5.32s (10 repetitions)
TOTAL code actions = 3.23s (10 repetitions)
TOTAL code actions after edit = 4.33s (10 repetitions)
Benchmark ghcide-bench: FINISH
```